+++ src/vm/allocator.cpp
@@ -2,11 +2,11 @@
 #include <stdlib.h>
 
 namespace gjs {
-    void* basic_malloc_allocator::allocate(u32 size) {
+    void* basic_malloc_allocator::allocate(u64 size) {
         return malloc(size);
     }
 
-    void* basic_malloc_allocator::reallocate(void* ptr, u32 size) {
+    void* basic_malloc_allocator::reallocate(void* ptr, u64 size) {
         return realloc(ptr, size);
     }
